create file with auxfiliar functions
